[jwies/review-scripted-only] review: split submission.test.ts under the max-lines budget
Absorbing the trial follow-ups put this file at 1094 lines against the
1000-line cap. Move the three follow-up suites (risks/patterns key
staging, the inline posting bar, the suppression verdict floor) into
submission-trial-followups.test.ts with local fixture copies, the same
split dispatch-gate-hardening.test.ts uses. Same 41 tests.
